refactor(dropDownItem): rename handleClick to onClick
